Fix transfer ratio type

+++ eosio.system/voting.cpp
@@ -60,7 +60,7 @@ namespace eosiosystem {
    //    }
    // }
 
-   void system_contract::regproducer( const account_name producer, const eosio::public_key& producer_key, asset maximum_supply, int64_t transfer_ratio, const std::string& url, uint16_t location ) {
+   void system_contract::regproducer( const account_name producer, const eosio::public_key& producer_key, asset maximum_supply, double transfer_ratio, const std::string& url, uint16_t location ) {
       eosio_assert( transfer_ratio > 0, "transfer ratio must be positive integer" );
       eosio_assert( url.size() < 512, "url too long" );
       eosio_assert( producer_key != eosio::public_key(), "public key should not be the default value" );
@@ -83,7 +83,7 @@ namespace eosiosystem {
       });
    }
 
-   void system_contract::updateproducer( const account_name producer, const eosio::public_key& producer_key, int64_t transfer_ratio, const std::string& url, uint16_t location ) {
+   void system_contract::updateproducer( const account_name producer, const eosio::public_key& producer_key, double transfer_ratio, const std::string& url, uint16_t location ) {
       eosio_assert( transfer_ratio > 0, "transfer ratio must be positive integer" );
       eosio_assert( url.size() < 512, "url too long" );
       eosio_assert( producer_key != eosio::public_key(), "public key should not be the default value" );
